Improve error message when not getting proper response from server
The response body could be empty - e.g. when the server can't be
connected - so showing only that as an error message might not always be
enough. The SoupMessage's reason_phrase might contain some information
about the error situation.

Now reason_phrase and the response body are shown in the error message
if the contain anything. If both are empty then a general "Unknown
error" is shown.